Debug - client publishes an event.

+++ src/proxy.cc
@@ -14,7 +14,8 @@ DEFINE_int32(port, 0, "Port that listens to incoming connections.");
 
 // Optional parameters.
 DEFINE_bool(on_disk, true, "Is on-disk storage enabled?");
-DEFINE_int64(mem_cap, 1 << 31, "Total size of in-memory buffer, 2 GB default.");
+DEFINE_int64(mem_cap, static_cast<int64_t>(1) << 31,
+    "Total size of in-memory buffer, 2 GB default.");
 DEFINE_int64(disk_chunk_size, 1 << 24,
     "Size of an on-disk data chunk, 16 MB default.");
 DEFINE_int32(gc_thread_count, 0, "Number of garbage collection threads.");
